initial menu tweak and settings file
no effective new functionality, but menu code restructured a bit to
facilitate the basic/advanced menu mode to come. Also have it loading
settings from a file, and creating the file on first run, though the
settings aren't used yet. Settings file goes to /etc if it exists,
otherwise to the directory edit+ is in.